update index page

+++ app/src/main/java/maimeng/yodian/app/client/android/model/skill/Theme.java
@@ -3,19 +3,17 @@ package maimeng.yodian.app.client.android.model.skill;
 import android.os.Parcel;
 import android.os.Parcelable;
 
-import java.io.Serializable;
-
 import maimeng.yodian.app.client.android.network.response.TypeData;
 
 /**
  * Created by android on 15-8-10.
  */
 public class Theme implements Parcelable, TypeData {
-    private long scid;
+    private int scid;
     private String name;
 
     protected Theme(Parcel in) {
-        scid = in.readLong();
+        scid = in.readInt();
         name = in.readString();
     }
 
@@ -31,7 +29,7 @@ public class Theme implements Parcelable, TypeData {
         }
     };
 
-    public long getScid() {
+    public int getScid() {
         return scid;
     }
 
@@ -39,7 +37,7 @@ public class Theme implements Parcelable, TypeData {
         return name;
     }
 
-    public void setScid(long scid) {
+    public void setScid(int scid) {
         this.scid = scid;
     }
 
@@ -58,7 +56,7 @@ public class Theme implements Parcelable, TypeData {
 
     @Override
     public void writeToParcel(Parcel dest, int flags) {
-        dest.writeLong(scid);
+        dest.writeInt(scid);
         dest.writeString(name);
     }
 }
